+++ veicoliDLL/Moto.cs
@@ -31,7 +31,7 @@ namespace veicoliDLLProject
         /// Costruttore "vero" della classe Moto, i parametri sono assegnati dinamicamente.
         /// </summary>
         /// <param>Spiegati in Veicolo.cs</param>
-        public Moto(string targa, string marca, string modello, string colore, int cilindrata, double potenzaKw, DateTime immatricolazione, bool isUsato, bool isKmZero, float kmPercorsi, string marcaSella, double prezzo, string imgPath = @".\img/genericMoto.jpg" /*Path dell'immagine, se omesso dalla dichiarazione si prende l'immagine generale*/)
+        public Moto(string targa, string marca, string modello, string colore, double cilindrata, double potenzaKw, DateTime immatricolazione, bool isUsato, bool isKmZero, float kmPercorsi, string marcaSella, double prezzo, string imgPath = @".\img/genericMoto.jpg" /*Path dell'immagine, se omesso dalla dichiarazione si prende l'immagine generale*/)
             : base(targa, marca, modello, colore, cilindrata, potenzaKw, immatricolazione, isUsato, isKmZero, kmPercorsi, prezzo, imgPath)
         {
             this.MarcaSella = marcaSella;
